compiles and links now, but as DOMifyining is not finished, it doesn't work at all but will soon
svn path=/trunk/koffice/; revision=26088

+++ kword/variable.cc
@@ -144,7 +144,7 @@ QDomElement KWPgNumVariable::save( QDomDocument& doc )
 
     QDomElement pgnum = doc.createElement( "PGNUM" );
     var.appendChild( pgnum );
-    pgnum.setAttribute( "value", pgNum );
+    pgnum.setAttribute( "value", (int)pgNum );
 
     return var;
 }
